requirements.txt.txt
numpy
pandas
matplotlib.pyplot
seaborn
os